+++ app/services/snapshot_service.rb
@@ -7,22 +7,25 @@ class SnapshotService
     @registry ||= Registry.new
   end
 
-  def initialize(paper, registry=SnapshotService.registry)
+  def initialize(paper, registry = SnapshotService.registry)
     @paper = paper
     @registry = registry
   end
 
   def snapshot!(*things_to_snapshot)
-    things_to_snapshot.flatten.each do |thing|
+    preview(*things_to_snapshot).each(&:save!)
+  end
+
+  def preview(*things_to_snapshot)
+    things_to_snapshot.flatten.map do |thing|
       serializer_klass = @registry.serializer_for(thing)
       json = serializer_klass.new(thing).as_json
-      Snapshot.create!(
+      Snapshot.new(
         source: thing,
         contents: json,
         paper: @paper,
         major_version: @paper.major_version,
-        minor_version: @paper.minor_version
-      )
+        minor_version: @paper.minor_version)
     end
   end
 end