+++ client/app/mixins/components/snapshot-named-computed-property.js
@@ -1,0 +1,8 @@
+export default function namedComputedProperty(name) {
+  return Ember.computed('snapshot.children.[]', function() {
+    let parts = this.get('snapshot.children');
+    return _.find(parts, function(part) {
+      return part.name === name;
+    }).value;
+  });
+};