+++ client/app/models/versioned-text.js
@@ -4,6 +4,8 @@ import DS from 'ember-data';
 export default DS.Model.extend({
   paper: DS.belongsTo('paper', { async: true }),
   text: DS.attr('string'),
-  created_at: DS.attr('date'),
-  version_string: DS.attr('string')
+  majorVersion: DS.attr(),
+  minorVersion: DS.attr(),
+  createdAt: DS.attr('date'),
+  versionString: DS.attr('string')
 });