+++ client/app/router.js
@@ -19,6 +19,8 @@ Router.map(function() {
       });
     });
 
+    this.route('versions', { path: '/versions' });
+
     this.route('workflow', function() {
       this.route('discussions', function() {
         this.route('new',  { path: '/new' });
@@ -28,6 +30,7 @@ Router.map(function() {
 
     this.route('task', { path: '/tasks' }, function() {
       this.route('index', { path: '/:task_id' });
+      this.route('version', { path: '/:task_id/version' });
     });
   });