+++ client/app/views/overlays/versioned_metadata.js
@@ -1,0 +1,5 @@
+import OverlayView from 'tahi/views/overlay';
+
+export default OverlayView.extend({
+  layoutName: 'layouts/overlay'
+});