+++ spec/support/pages/overlays/versioned_metadata_overlay.rb
@@ -1,0 +1,5 @@
+class VersionedMetadataOverlay < CardOverlay
+  def expect_version(version)
+    expect(page).to have_content(version)
+  end
+end